dataflow/logging: Rework the reachability logging
The change does two things: It enables early data consolidation and tries
to make row packing more efficient. Experiments show that the reachability
analysis spends a large amount of time in creating `Row`s. Due to the
current approach of keying the data before it's arranged, the row was first
constructed, then unpacked and repacked, and finally shipped with the
original row. This change passes the key expression to a closure, which
then constructs the dataflow and directly produces the correctly keyed
row.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>