expr: resolve sql func output type with generics
The previous syntactic method of converting the captured return type
into a ScalarType relied the types being imported in the generated
namespace of the macro. This isn't true for many custom types.

This patch uses a generic function instead that extracts the return type
and converts it to a ScalarType without relying on this syntactic
property.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>