sql-parser: remove dependency on repr
repr and sql-parser share a single, tiny data structure, but otherwise
don't know about each other. Duplicate this type so the crates don't
depend on each other. For users of just sql-parser, this reduces total
dependencies (not really a problem for materialize directly). Materialize
gets the benefit that some of our larger crates are no longer coupled,
at the cost of a small transform function.