expr: allow any parameter type for explicit sqlfunc!
If all the required attributes of the function are specified we don't
need to enforce an Option<T> structure to the parameter type. The only
requiremenent is that the parameter type can be converted into a Datum.

This change allows defining functions using the sqlfunc! macro that
accept any Datum and produce a concrete value, e.g is_null that goes
from any datum -> bool.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>