dataflow/logging: Pre-arrange Timely logs before mapping to Row
Previously, logs were demuxed, converted to rows and arranged. The default
implementation for arranging Rows is not as fast as arranging the raw data.
This change adds an arrangement for streams that are known to have a high
volume of data which is not reduced be the worker-local pre-aggregation
within the demux operator.

In testing, it reduces the time to process logs in exchange for a slight
increase in memory compared to the parent change.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>