persist: turn Persist::persist() operator into modular persist()/seal() operators
This turns the monolithic persist() operator, which was responsible for
both persisting and replaying persisted data into a simpler persist()
operator that only persists. We also add a seal() operator that can seal
a collection based on the advancing frontier.

These are also more efficient (less blocking) because we don't recv(),
that is block, on every write. Instead, we now buffer write futures and
wait on them when the frontier advances.

We will use these operators in the implementation of
persistent_upsert(), and when implementing timestamp binding persistence
for sources.