persist: add conditional_seal() operator
Passes through each element of the stream and seals the given primary
and condition collections, respectively, when their frontier advances.
The primary collection is only sealed up to a time `t` when the
condition collection has also been sealed up to `t`.

This does not wait for the seals to succeed before passing through the
data. We do, however, wait for the seals to be successful before
allowing the frontier to advance. In other words, this operator is
holding on to capabilities as long as seals corresponding to their
timestamp are not done.