persist: allow multiple workers to together concurrently read a snapshot
Previously, a snapshot had to be read entirely in one place, limiting
the concurrency with which this can be done. This will soon become a
scaling bottleneck, so get ahead of it. This commit adds the ability to
break a snapshot into a given number of roughly equal pieces. At the
same time, it makes the iteration itself more idiomatic by implementing
the Iterator trait. (The previous iteration mechanism was my fault and
an ultimately doomed premature performance optimization.)

In the process, we lost the ability to associate a Codec::decode error
with an individual key or value, but at the moment it doesn't seem
particularly important to maintain that. We can always add it back in
later.

This is part 1 of 2. The next part will actually use this new ability in
persist's replay operator to give each worker an equal chunk of the
work.