persist: begin an abstraction for maintenance in a background threadpool
This uses a threadpool for the cpu-intensive and io work, specifically
the `tokio::runtime::Runtime` used elsewhere in Materialize. Maintenance
work is done asynchronously and a response is returned as a Future.

To break this work into steps, the current Indexed runtime simply blocks
on this Future in the old call site. A followup will break the request
sending and response handling into separate places in Indexed, which is
non-trivial. That work will also have to ensure we don't send out the
same compaction requests repeatedly while waiting for a response, which
will require tracking a bit of state we hadn't before.

The implementation of compact_trace_blocking in this commit is identical
to the previous one in `Trace::merge`. All changes are made in the next
commit.