materialized: hook up s3 persist storage to command line flags
This allows selection of S3 as the storage used by persistence with a
new `--persist-s3-bucket` cli flag to `materialized`. The optional
`--persist-s3-prefix` flag can also be used to prepend to each of the
object keys. If `--persist-s3-bucket` is unset and persistence is
enabled, we fall back to using file storage in the data directory. S3
credentials and region must be available in the process or environment.

`--persist-s3-bucket` is only allowed with `--experimental`.