coord: restrict table mutation selections to other user tables
To ensure serializability, we need to ensure ReadThenWritePlans
never refer to objects that can receive writes between the plan's
read and its write. The only objects for which we can make this
guarantee are user-defined tables, which we lock for the length
of the operation.

Note that this patch also fixes a bug introduced in b99aec2,
which let UPDATE/DELETE use subqueries, which introduced
potential serializability violations.