pgwire: don't chain row batches
The `.chain()` call in `send_rows()` could cause a stack overflow if there
were a high number of batches being processed through `TAIL`. This was
occurring because each .chain would return a new Chain object, nesting
some other Chain. When there were enough, calling .next on them would
fill up the stack while they tried to get to the actual row batch.

This was part of the design because we needed some way to requeue the
current partial batch. Fix the problem by creating a specific location
for an existing batch to go. We can check it first, and only fetch the
next batch if it doesn't exist. When we have a remaining partial batch
it can be inserted back, removing the call to chain completely.

Previously RowBatchStream was a Box dyn because the object type could
either be an UnboundedReceiver or a Chain. Since it is now never a chain,
we can also take this opportunity to de-box-and-dyn the type and make
it concrete.

Fixes #8270