sql: plan VALUES clauses via a table function
Introduce TableFunc::Wrap, which takes N expressions as input and chunks
the expressions into rows of a specified width. Then teach the SQL
planner to plan VALUES clauses via this table function, rather than
unioning a bunch of Map expressions together.

The impetus is creating simpler plans post decorrelation:

    materialize=> explain decorrelated plan for select * from (values(1), (2));
      Decorrelated Plan
    ----------------------
     %0 = Let l0 =       +
     | Constant ()       +
                         +
     %1 =                +
     | Get %0 (l0)       +
     | FlatMap wrap1(1, 2)+

    materialize=> explain raw plan for select * from (values (1), (2), (1+1));
                Raw Plan
    ---------------------------------
     %0 =                           +
     | CallTable wrap1(1, 2, (1 + 1))+

This is an alternative to #8467.