dataflow/logging: Activate on data
Currently, the logging dataflows activate once every introspection
interval to process data. This can cause a significant buildup of data
before it is processed.

In addition to activating the replay operator periodically, also
activate it once a sufficient amount of data has been accumulated.
Ideally, the data processing should be activated as soon as some data is
available, but this could cause a situation where activating causes
additional data to be produced.

For this reason, the dataflow activates after 32 batches have been
published by the logging infrastructure. This seems to be a reasonable
trade-off between memory (currently at most 8KiB per batch) and latency.

The time-based activation is still required in case there is no data for
a specific dataflow. In this case, we still need to advance the clock,
which the time-based activation takes care of.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>