dataflow: indent match body correctly
For some reason rustfmt thought that the previous indentation was the
correct one. Weirdly, after manually fixing it, re-running `cargo fmt`
does *not* immediatelly revert back to the broken state but formats the
whole file correctly. Then if you run rustfmt a second time on the exact
same input it outputed a second ago, it reverts back to the broken
syntax.

That's probably an indication of a rustfmt glitch but I can't be
bothered to make a minimal testcase. A skip directive is added until the
code is refactored or rustfmt gets fixed

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>