persist: in conditional_seal(), eagerly persist up to frontier
Before, we were using a Notificator to notify us when the frontier
passed a capability. This mean that we would wait longer than we have
to: a capability with timestamp t is only considered "done" when the
frontier is at least t+1, meaning we were using at least one extra cycle
of waiting.

Now, we seal up directly to the combined input frontier.

This also means that we have to adapt the test, because we never seal
timestamp `0`, which is considered to be sealed from the beginning.