reduce clones by taking owned args
The affected functions were unconditionally cloning arguments that were
references. This can lead to waste if the caller would otherwise be
happy to move ownership of the value to the affected function.

This patch makes all of these functions require an owned arguments and
push the responsibility of clone vs move to the caller. In a few places
this has resulted in less clones overall.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>