persist: add get_description()
This allows getting a _differential_ `Description` of a persisted
stream.

This will be useful (needed even) when trying to determine a common
restore timestamp for multiple entangled streams backing a source and
when determining if the `since` of a stream is adequate for what the
rest of the dataflow needs.