dataflow: make decoders oblivious of position metadata
The decoding stage is responsible for transforming a timely stream of
`SourceOutput` messages into a timely stream of `DecodeResult`.
Previously, depending on the `push_metadata` flag, the format decoders
would push the position metadata into the decoded row. However,
`DecodeResult`s have a `position` field that carries the position
metadata anyway.

In this patch format decoder implementations are made oblivious of the
`position`[1] and `push_metadata` information and never produce this
information in the output row.

Then, the downstream envelope stage of transforming the `DecodeResult`
stream into a pure `Row` stream can choose whether or not it makes sense
for the envelope to append the position metadata into the output `Row`.

[1] The avro format decoder still needs access to position information
because it's intermingled with debezium logic. The property of never
pushing the metadata into the output Row still holds though, which is
sufficient for this patch. Once the debezium logic is fully implemented
on top of Rows the Avro decoder will be oblivious of position
information.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>