persist,dataflow: for persistent tables, make operator construction deterministic
Before, we were potentially constructing differently shaped operator
graphs on each worker, based on the result of the `create_or_load()`
call.

Now, we pass that result to `persistent_source()` and handle that error
internally. Meaning we don't construct a separate `Stream` of errors but
instead let the operator emit the error on its output.

Fixes #8800