add a timeout to ccsr requests
Currently all ccsr requests get the reqwest client's default infinite
timeout. This can cause materialize to hang forever if it's eg. pointed
at a slow/malicious server, a server that isn't actually a schema
registry, etc.

This was noticed because
https://github.com/MaterializeInc/materialize/pull/8743 added a test
where the client points at something that isn't an actual schema
registry. When it gets a response back (even a 4xx or 5xx), the test
succeeds. Redpanda in nightlies does not throw an error for that test
and ended up in progress for hours until the CI run itself timed out.