persist: in conditional_seal() operator, don't block on seal futures
Before, we would synchronously block on the seal future using recv().
This meant that we block the timely worker thread far longer than we
need to.

Now, we convert it into a "real" Future and poll it asynchronously.

Additionally, we simplify capability handling/downgrading by switching
to a `CapabilitySet`.

The test needed to be adapted for this, because the results are not
deterministic anymore. Now, we assert on the ordering properties that we
want to have.

With this change and the change to make persist() non-blocking, we get
these benchmark results for `avro-upsert` on a Ryzen 3700X with 64GB of
ram, with persistence enabled:

./bin/mzbench --size medium --benchmark-id hello --num-measurements 1 avro-upsert
git_revision,num_workers,iteration,seconds_taken,rows_per_second
None,7,0,25,800000
None,5,0,34,588235
None,4,0,43,465116
None,2,0,84,238095

This is the same as the result without persistence. However,
`avro-upsert` is not correct because it only measures how many messages
we ingested from Kafka, without measuring the time to go through
persistence. That is, we don't measure when the data would be available
to query.

Closes #8812