persist: add list_keys to Blob (#8684)
This commit adds a method that shows users the set of keys currently set in the
map. We will use this method in the future to correctly delete all non-lock keys
from the underlying storage (e.g. the parent directory or the containing bucket
and matching prefix) when we detect a version mismatch.

We have to delete all keys, rather than more simply deleting the underlying storage
because the S3 API requires that a bucket be empty before it can be deleted.
See: https://docs.rs/rusoto_s3/0.47.0/rusoto_s3/struct.S3Client.html#method.delete_bucket

We can also use this method in the future to detect and delete orphaned keys that
are no longer in use.