transform: Reuse knowledge_stack allocation
The ColumnKnowledge implementation requires a stack to determine its
DatumKnowledge. The stack is a vector that starts empty and might be
reallocated several times while optimizing the expression.

With this change, the transformation re-uses the same allocation, which
saves new allocations and re-allocations. There is no downside besides
increased code complexity because the maximum amount of memory consumed
by the knowledge_stack remains the same, and the memory is only retained
for the duration of transform call.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>