dataflow,persist: wait for updates to be persisted before forwarding from source
This brings the behaviour on line with what we wanted to have for the
initial implementation. It's not wrong to forward updates to dataflow
early, while holding on to the frontier.

This approach is the most conservative option, where we absolutely make
sure that no-one can see updates downstream until we persisted them.