persist: introduce ingest-side workers in nemesis
Nemesis recently got the ability to run multiple workers on the
dataflowd-side, as well as running dataflowd concurrently with
operations on the ingestd-side. Continue adding more concurrency
coverage to nemesis, by introducing the idea of a worker to the
ingestd-side too. Only one of these workers is used initially (getting
exactly the previous behavior) because the Validator changes necessary
to make sense of the increased concurrency are substantial and this lets
us split up the work into more manageable PR sizes.

For easy of understanding (and review) of this commit, nothing is
reordered and nothing is renamed, resulting in a very awkward set of
impl blocks. These cleanups are done in the next commit, which makes no
behavior changes.