persist: give each mutating request a SeqNo
Persist can be described as a state machine, with input state changes
being sequenced into a log. As an initial step in the direction of
structuring the implementation in those terms, give each request that
mutates state a unique SeqNo corresponding to the order it was applied.

This ordering also includes requests that were sequenced and applied to
the persist state machine, but that application was deterministically
made into a no-op because it was contextually invalid (a write or seal
at a sealed timestamp, an allow_compactions at an unsealed timestamp,
etc).

Read-only requests are assigned the SeqNo of a write, indicating that
all mutating requests up to and including that one are reflected in the
read state.

The immediate motivation for this is that it allows for much easier
post-hoc reasoning about behavior in the nemesis testing, but it also
will also tee us up for upcoming incremental meta work.